+++ landingzones/management/commands/inactivezones.py
@@ -11,14 +11,14 @@ from landingzones.models import LandingZone
 
 
 def get_inactive_zones(weeks=2):
-    """Return list of landing zone modified old than n weeks."""
+    """Return landing zones last modified over n weeks ago"""
     return LandingZone.objects.filter(
         date_modified__lte=localtime() - timedelta(weeks=weeks)
     ).exclude(status__in=('DELETED', 'MOVED'))
 
 
 def get_output(zones, irods_backend):
-    """Return list of enriched inactive landing zones."""
+    """Return list of inactive landing zone details"""
     lines = []
     for zone in zones:
         path = irods_backend.get_path(zone)
@@ -39,9 +39,9 @@ def get_output(zones, irods_backend):
 
 
 class Command(BaseCommand):
-    """Command to find landingzones last modified more than two weeks ago."""
+    """Command to find landing zones last modified more than two weeks ago"""
 
-    help = 'Find landingzones last modified more than two weeks ago.'
+    help = 'Returns list of landing zones last modified over two weeks ago'
 
     def handle(self, *args, **options):
         irods_backend = get_backend_api('omics_irods')